Updated dotfiles to not be at root directory
I got tired of all the downsides of having a root git directory for
dotfiles, so I migrated everything over to be within a dotfiles
directory.

This creates some symlinks for the main config files:
- .bash_profile (Mac weird thing where it uses it for each shell)
- .bashrc
- .vimrc

Also created the `init.sh` script to be included here instead of relying
on the gist from github.